Trigger rates after search WIP
